update student info, update dance info
Also reorganized the forms a little bit.
