+++ src/components/Users/new.js
@@ -1,9 +1,9 @@
 // @flow
-// src/components/NewStudentForm/index.js
+// src/components/Users/index.js
 import React, { PureComponent } from "react";
 import { Button, Form, FormGroup, Label, Input } from 'reactstrap';
 import type { User } from "../../types.js";
-import { addNewUser, signInUser } from "../../lib/api.js";
+import { addNewUser } from "../../lib/api.js";
 
 type State = User;
 
@@ -36,7 +36,9 @@ class NewUserForm extends PureComponent<Props, State> {
   onSubmit = (event: any) => {
     event.preventDefault();
     // Validate form
-    addNewUser(this.state);
+    addNewUser(this.state).then(function() {
+      window.location.href = "/";
+    });
     // Clear the form
     this.clearForm();
   };
@@ -63,72 +65,9 @@ class NewUserForm extends PureComponent<Props, State> {
           <br></br>
           <Button type="submit" value="Submit">Submit</Button>
         </Form>
-
-        <br></br>
-        <div>
-        {/*<code>{JSON.stringify(this.state)}</code>*/}
-        </div>
-        <br></br>
-
       </div>
     );
   }
 }
 
-class SignInUserForm extends PureComponent<Props, State> {
-  state: State = {
-    email: "",
-    password: ""
-  };
-
-  onChange = (event: any) => {
-    const { name, value } = event.target;
-    this.setState({
-      [name]: value
-    });
-  };
-
-  clearForm() {
-    this.setState({
-      email: "",
-      password: ""
-    });
-  };
-
-  onSubmit = (event: any) => {
-    event.preventDefault();
-    // Validate form
-    signInUser(this.state);
-    // Clear the form
-    this.clearForm();
-  };
-
-  render() {
-
-    return (
-      <div className="App">
-        <h1>Sign In</h1>
-        <p>Admins only</p>
-        <Form onSubmit={this.onSubmit}>
-          <FormGroup>
-            <Label form="email" type="email">Email</Label><Input placeholder="me@example.com" onChange={this.onChange} value={this.state.email} type="email" id="email" name="email" />
-          </FormGroup>
-          <FormGroup>
-            <Label form="password" type="password">Password</Label><Input onChange={this.onChange} value={this.state.password} type="password" id="password" name="password" />
-          </FormGroup>
-          <br></br>
-          <Button type="submit" value="Submit">Submit</Button>
-        </Form>
-
-        <br></br>
-        <div>
-        {/*<code>{JSON.stringify(this.state)}</code>*/}
-        </div>
-        <br></br>
-
-      </div>
-    );
-  }
-}
-
-export { NewUserForm, SignInUserForm };
+export default NewUserForm;
