+++ src/routes.js
@@ -1,21 +1,24 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom'
 import Home from './components/Home';
-import { NewUserForm, SignInUserForm } from './components/User';
-import NewStudent from './components/NewStudentForm';
-import ReturningStudent from './components/ReturningStudentForm';
+import SignInUserForm from './components/Users';
+import NewUserForm from './components/Users/new.js';
+import NewStudentPage from './components/NewStudentForm';
+import ReturningStudentPage from './components/ReturningStudentForm';
 import Dance from './components/DanceForm';
-import DanceCheckinForm from './components/DanceCheckinForm';
+import DanceCheckinPage from './components/DanceCheckinForm';
+import StudentPage from './components/Students';
 
 const Routes = (props) => (
   <Switch {...props}>
     <Route exact path="/" component={Home} />
     <Route path="/new-user" component={NewUserForm} />
     <Route path="/signin" component={SignInUserForm} />
-    <Route path="/new-student" component={NewStudent} />
-    <Route path="/returning" component={ReturningStudent} />
+    <Route path="/new-student" component={NewStudentPage} />
+    <Route path="/returning" component={ReturningStudentPage} />
     <Route path="/dance" component={Dance} />
-    <Route path="/dance-checkin" component={DanceCheckinForm} />
+    <Route path="/dance-checkin" component={DanceCheckinPage} />
+    <Route path="/student" component={StudentPage} />
   </Switch>
 );
